requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.0
httpx==0.27.2
streamlit==1.28.0
requests==2.31.0
pydantic<2.0.0
python-dotenv==1.0.0

